Ask for database when adding allergies, not in the main allergies activity
